# BookstoreSwagger.yml
---
openapi: "3.0.0"
info:
  title: "Book API"
  description: "API для управления книгами"
  version: "1.0.0"
servers:
  - url: "http://127.0.0.1:8080"
paths:
  /book:
    get:
      summary: "Get all books"
      parameters:
        - name: "limit"
          in: "query"
          description: "The maximum number of books to return"
          required: "false"
          type: "integer"
        - name: "title"
          in: "query"
          description: "Filter books by title"
          required: "false"
          type: "string"
      responses:
        200:
          description: "Successful operation"
          schema:
            type: "array"
            items:
              $ref: "#/definitions/Book"
        400:
          description: "Invalid limit parameter"
        500:
          description: "Internal server error"
    post:
      summary: "Add a new book"
      parameters:
        - name: "book"
          in: "body"
          description: "The book object to add"
          required: "true"
          schema:
            $ref: "#/definitions/Book"
      responses:
        200:
          description: "Successful operation"
          schema:
            $ref: "#/definitions/Book"
        400:
          description: "Invalid request body"
        500:
          description: "Internal server error"
  /book/{id}:
    get:
      summary: "Get a book by ID"
      parameters:
        - name: "id"
          in: "path"
          description: "The ID of the book to retrieve"
          required: "true"
          type: "integer"
      responses:
        200:
          description: "Successful operation"
          schema:
            $ref: "#/definitions/Book"
        400:
          description: "Invalid ID parameter"
        404:
          description: "Book not found"
        500:
          description: "Internal server error"
    put:
      summary: "Update a book by ID"
      parameters:
        - name: "id"
          in: "path"
          description: "The ID of the book to update"
          required: "true"
          type: "integer"
        - name: "book"
          in: "body"
          description: "The updated book object"
          required: "true"
          schema:
            $ref: "#/definitions/Book"
      responses:
        200:
          description: "Successful operation"
        400:
          description: "Invalid ID parameter or request body"
        500:
          description: "Internal server error"
    delete:
      summary: "Delete a book by ID"
      parameters:
        - name: "id"
          in: "path"
          description: "The ID of the book to delete"
          required: "true"
          type: "integer"
      responses:
        204:
          description: "Successful operation"
        400:
          description: "Invalid ID parameter"
        404:
          description: "Book not found"
        500:
          description: "Internal server error"
  /books:
    get:
      summary: "Get all books"
      responses:
        200:
          description: "Successful operation"
          schema:
            type: "array"
            items:
              $ref: "#/definitions/Book"
definitions:
  Book:
    type: "object"
    properties:
      id:
        type: "integer"
      title:
        type: "string"
      authors:
        type: "array"
        items:
          type: "string"
      year:
        type: "integer"